Created particle filter
Particle - a single particle with a state vector, weight (or likelihood), and timestamp.
ParticleState - An equivalent to the likes of GaussianState where the state of an object is represented by a cluster of particles. State mean and covariance are the mean and covariance of the entire distribution (assumes all particles have equal likelihood)
Resampler - a method of resampling a distribution of weighted particles. A SystematicResampler is included