Tests fixed and updated
Noise diffusion coefficients in the tests are now passed in as (n, 1) array instead of (n, singleton array). Test for Singer Model is updated to correctly test the case where time difference is large and the full covariance matrix should be used instead of the simplified one that only calculates the leading terms.