Add parallel Kalman filter for particle flow filters
This requires being able to fix the covariance of the ParticleState to
that of the Kalman filter, so this has also been added to the
ParticleState.

A change has also been made to the ParticleUpdater to copy particles to
ensure the original predicted state isn't overwritten, and use of
resampler is now optional.