JPDA for asynchronous measurements
The prediction and measurement prediction for every hypothesis in the multihypothesis were originally set as for the missed detection state. This causes spurious results at the update step if the measurements in the multihypothesis are asynchronous (e.g. grouped by a time-windowing feeder class), and the sensor platform is moving. In this case, the updated state is calculated using the sensor's translation offset at the missed detection timestamp (i.e. the latest timestamp in the window), but for a position measurement taken at the measurement timestamp (which may be earlier in the window).
This fix attaches the prediction/measurement prediction calculated for the measurement to the relevant SingleHypothesis, such that the update is calculated using the correct sensor offset for the measurement timestamp.