Use Measures in track to track association and OSPA metrics.
This allows more than Euclidean distance to be used, and use of Measures
mapping to select which dimensions to compare, removing need for
measurement models, which were counter-intuitive.

I think there may be a case where the dimensions of track and truth
don't match in order to use mapping, but suggest better approach would
be to extend Measures to allow per state mapping.