fix a test for MovingMovable
The change made in StateMutableSequence in the previous commit affects the error raised when MovingMovable is instantiated without specifying the mandatory argument "states". This change has made a test for MovingMovable fail, although I think for most users, the change itself is harmless. Therefore, I fix the test to make it pass.