Check for State type more explicitly with StateMutableSequence
There is some inconsistent behaviour between Python versions, whether
running code in debug or tests run; that effect the isinstance check
with States that may have properties similar to StateMutableSequnce
type.

This particularly seems to effect ASDState, and multi-model particle
states. This extra check now ensure that anything that in itself is
defined as a State type is mistaken.